+++ src/libfetchers/memory-input-accessor.hh
@@ -1,4 +1,5 @@
 #include "input-accessor.hh"
+#include "source-path.hh"
 
 namespace nix {
